pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-pr-reviewer"
version = "0.1.0"
description = "AI-powered PR review agent with Azure OpenAI GPT-5.4"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyjwt[crypto]>=2.10.0",
    "pyyaml>=6.0.2",
    "structlog>=24.4.0",
    "azure-identity>=1.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-PyYAML>=6.0.12",
    "respx>=0.22.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/reviewer"]

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "S",    # flake8-bandit
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
    "ASYNC",# flake8-async
]
ignore = [
    "S101",  # assert allowed in tests
    "B008",  # function call in default arg (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]
"tests/queue/__init__.py" = ["A005"]
"tests/validate.py" = ["T201"]
"src/reviewer/middleware/*.py" = ["TC001", "TC002", "TC003"]
"src/reviewer/server/*.py" = ["TC001", "TC002"]
"src/reviewer/main.py" = ["TC001", "TC003"]
"src/reviewer/reviewer/agent.py" = ["TC003"]
"src/reviewer/reviewer/pipeline.py" = ["TC003"]
"src/reviewer/reviewer/tools.py" = ["TC003"]

[tool.ruff.lint.isort]
known-first-party = ["reviewer"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_decorators = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/reviewer"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "\\.\\.\\.",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
